TLS/SSL headers can be decoded now.
This is not decryption, just reading the wire format of the TLS
protocol. For now, it is just the TLSPlaintext structure, and even that
just support for ClientHello content.

It is intended to be used to verify properties of TLS connections that
otherwise may be difficult or inaccessible from the SSLEngine for
whatever reason.