Make it so we can pass the TrustManagers down to the handshake exception handler. All of this is in the name of currying things along from SSLContext into the SSLSocket so that we can know the TrustManager when a handshake fails.
Some big automated refactoring:
* AbstractXYZProxy -> XYZProxy
* Custom proxy subclasses are now TealessXYZ

Why so many proxies? Here is why:
I tracked things through the JDK, and I don't think we can reasonably access this without passing it explicitly. The call chain looks something like this:
SSLContext.getSocketFactory() -> SSLContextSpi.engineSocketFactory() -> SSLSocketFactoryImpl.createSocket(context, ...) -> SSLSocketImpl.startHandshake() -> ClientHandshaker

Most of these things (the Impl classes, etc) are packaged in the `sun.` namespace. ClientHandshaker has handle on the SSLContextImpl and asks `.getX509TrustManager()` which returns an X509ExtendedTrustManager which .checkTrusted(chain, authType, socket, isClient) is called on.

All of this is to say we want to be able to tell an SSLContext that we want to be able to know the TrustManager when a handshake failure happens. Because ClientHandshaker+SSLSocketImpl are internal things to the jdk, I'd prefer try to use them.