Add bad_certificate alert handling.
Only flip buffers when we are reading the oplog. This is necessary
because when bad_certificate happens, we may get two exceptions: one
from the client writer which gets `Broken Pipe` and one from the client
reader which gets the SSLHandshakeException. This helps us hopefully
only read the log once. Long term, I'll probably want to handle this
more gracefully to allow log reading multiple times.